refactor: reorganize to modern Go layout (cmd/ + internal/)
Move editor core from root-level `package main` into `internal/editor/`,
treepane plugin into `internal/plugins/treepane/`, and the binary entry
point into `cmd/nistru/`. Extract an `editor.Run(path)` entry point so
`cmd/nistru/main.go` is a thin ~15-LOC wrapper. Add a `repoRoot(t)`
helper to resolve `examples/plugins/...` from the new test package
location. Root goes from ~25 top-level files (11 Go sources) to 18
top-level files (0 Go sources).

Public packages (`plugin/`, `sdk/plugsdk/`, `examples/`) unchanged.

BREAKING CHANGE: `go install github.com/savimcio/nistru@latest` no
longer works. Install via `go install github.com/savimcio/nistru/cmd/nistru@latest`
instead.
