+++ .gitignore
@@ -1,5 +1,6 @@
 .claude/plans/
 /nistru
+cmd/nistru/nistru
 /examples/plugins/hello-world/hello-world
 /examples/plugins/gofmt/gofmt-plugin
 *.swp
