+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Reorganize repo to modern Go layout: editor core moves to `internal/editor/`, binary entry point to `cmd/nistru/`, first-party plugins to `internal/plugins/`. Public packages (`plugin/`, `sdk/plugsdk/`) unchanged.
+
+### Breaking
+
+- **`go install github.com/savimcio/nistru@latest` no longer works; install via `go install github.com/savimcio/nistru/cmd/nistru@latest` instead.**
+
 ## [0.1.0] — 2026-04-21
 
 Initial public release.
