+++ internal/editor/autosave.go
@@ -1,4 +1,4 @@
-package main
+package editor
 
 import (
 	"fmt"
