+++ internal/editor/autosave_test.go
@@ -1,4 +1,4 @@
-package main
+package editor
 
 import (
 	"os"
