+++ internal/editor/palette.go
@@ -1,4 +1,4 @@
-package main
+package editor
 
 import (
 	"sort"
