+++ internal/editor/testdata/golden/empty_120x40.txt
