+++ internal/editor/testdata/golden/empty_40x10.txt
