+++ internal/editor/testdata/golden/empty_80x24.txt
