+++ internal/editor/testdata/golden/long_path_truncation_40x10.txt
