+++ internal/editor/testdata/golden/open_go_120x40.txt
