+++ internal/editor/testdata/golden/open_go_80x24.txt
