+++ internal/editor/testdata/golden/palette_filtered_80x24.txt
