+++ internal/editor/testdata/golden/palette_open_80x24.txt
