+++ internal/editor/testdata/golden/status_segments_80x24.txt
