+++ internal/editor/testdata/golden/treepane_expanded_80x24.txt
