+++ internal/plugins/treepane/treepane.go
