+++ internal/plugins/treepane/treepane_test.go
